+++ traffic_router/core/src/main/scripts/postinstall.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,12 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+keytool=$(dirname $(readlink -f $(which java)))/keytool
+cd /opt/traffic_router/conf
 if [ -f /opt/traffic_router/conf/*.crt ]; then
-	cd /opt/traffic_router/conf
 	for file in *.crt; do
 		alias=$(echo $file |sed -e 's/.crt//g' |tr [:upper:] [:lower:])
 		cacerts=$(/bin/find $(dirname $(readlink -f $(which java)))/.. -name cacerts)
-		keytool=$(dirname $(readlink -f $(which java)))/keytool
 		$keytool -list -alias $alias -keystore $cacerts -storepass changeit -noprompt > /dev/null
 
 		if [ $? -ne 0 ]; then
@@ -26,6 +27,17 @@ if [ -f /opt/traffic_router/conf/*.crt ]; then
 	done
 fi
 
+
+echo -e "
+cd /opt/traffic_router/conf
+
+keytool=\$(dirname \$(readlink -f \$(which java)))/keytool
+
+if [ ! -f /opt/traffic_router/conf/keyStore.jks ]; then \n
+    \$keytool -genkeypair -v -alias \$(hostname -f) -dname \"CN=\$(hostname -f), OU=APIDefault, O=Apache Traffic Control, L=Denver, ST=Colorado, C=US\" -keystore \$(pwd)/keyStore.jks -storepass changeit -keyalg RSA -ext KeyUsage=\"digitalSignature,keyEncipherment,keyCertSign\" -ext BasicConstraints:\"critical=ca:true\" -storetype JKS -validity 3650
+    \$keytool -exportcert -v -alias \$(hostname -f) -file \$(hostname -f).crt -keypass changeit -storepass changeit -keystore \$(pwd)/keyStore.jks -rfc
+fi" >> generatingCerts.sh
+chmod 755 generatingCerts.sh
 
 echo "Traffic Router installed successfully."